Rename middleware files to kebab-case
